chore: begin refactor
