+++ gradle.properties
@@ -1,20 +1,21 @@
 # Gradle performance improvements
 org.gradle.jvmargs=-Xmx2048M -XX:MaxMetaspaceSize=512m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8 -XX:+UseParallelGC
 org.gradle.vfs.watch=true
-kapt.use.worker.api=true
-kapt.include.compile.classpath=false
 org.gradle.parallel=true
 org.gradle.caching=true
 org.gradle.warning.mode=all
+kapt.use.worker.api=true
+kapt.include.compile.classpath=false
+
 # Fabric Properties
-# check these on https://modmuss50.me/fabric.html
-minecraft_version=1.19.4
-yarn_mappings=1.19.4+build.2
-loader_version=0.14.19
+    # check these on https://modmuss50.me/fabric.html
+    minecraft_version=1.19.4
+    yarn_mappings=1.19.4+build.2
+    loader_version=0.14.19
 # Mod Properties
-mod_version=1.19.4-beta3
-maven_group=gaucho-matrero.altoclef
-archives_base_name=altoclef
+    mod_version=1.19.4-beta4
+    maven_group=dev.toastmc
+    archives_base_name=bartione-plus
 # Dependencies
-# check this on https://modmuss50.me/fabric.html
-fabric_version=0.77.0+1.19.4
+    # check this on https://modmuss50.me/fabric.html
+    fabric_version=0.77.0+1.19.4
