+++ settings.gradle
@@ -7,3 +7,4 @@ pluginManagement {
         gradlePluginPortal()
     }
 }
+rootProject.name = 'Baritone+'
