+++ src/main/java/adris/altoclef/brainWIP/utils/WorldState.java
@@ -1,4 +1,4 @@
-package adris.altoclef.baritone.brain.utils;
+package adris.altoclef.brainWIP.utils;
 
 import adris.altoclef.AltoClef;
 import com.google.common.collect.ImmutableList;
