+++ src/main/java/adris/altoclef/chains/UserTaskChain.java
@@ -2,7 +2,7 @@ package adris.altoclef.chains;
 
 import adris.altoclef.AltoClef;
 import adris.altoclef.Debug;
-import adris.altoclef.baritone.brain.BaritoneBrain;
+import adris.altoclef.brainWIP.BaritoneBrain;
 import adris.altoclef.eventbus.EventBus;
 import adris.altoclef.eventbus.events.TaskFinishedEvent;
 import adris.altoclef.tasksystem.Task;
