+++ src/main/resources/baritone-plus.mixins.json
@@ -6,28 +6,26 @@
   "mixins": [
     "AbstractFurnaceScreenHandlerAccessor",
     "BlockModifiedByPlayerMixin",
+    "ChatInputMixin",
+    "ChatReadMixin",
     "ClientBlockBreakAccessor",
+    "ClientBlockBreakMixin",
     "ClientConnectionAccessor",
     "ClientInteractWithBlockMixin",
+    "ClientOpenScreenMixin",
     "ClientPlayerInteractionAccessor",
+    "ClientTickMixin",
     "ClientUIMixin",
     "DeathScreenAccessor",
     "EntityAccessor",
-    "GameOverlayMixin",
-    "PersistentProjectileEntityAccessor",
-    "SlotClickMixin",
-    "WorldBlockModifiedMixin"
-  ],
-  "client": [
-    "ChatInputMixin",
-    "ChatReadMixin",
-    "ClientBlockBreakMixin",
-    "ClientOpenScreenMixin",
-    "ClientTickMixin",
     "EntryMixin",
+    "GameOverlayMixin",
     "LoadChunkMixin",
     "MixinLocalPlayer",
+    "PersistentProjectileEntityAccessor",
     "PlayerCollidesWithEntityMixin",
+    "SlotClickMixin",
+    "WorldBlockModifiedMixin",
     "WorldRendererMixin"
   ],
   "injectors": {
