+++ src/main/resources/fabric.mod.json
@@ -1,20 +1,20 @@
 {
   "schemaVersion": 1,
-  "id": "altoclef",
-  "version": "1.19.4-beta3",
-  "name": "Alto Clef",
+  "id": "baritone-plus",
+  "version": "${version}",
+  "name": "Baritone+",
   "description": "Plays Block Game",
   "authors": [
-    "TacoTechnica",
-    "James Green",
-    "Marvion Kirito"
+    "RemainingToast",
+    "Baritone",
+    "AltoClef"
   ],
   "contact": {
-    "homepage": "https://fabricc.net/",
-    "sources": "https://github.com/FabricMC/fabric-example-mod"
+    "homepage": "https://toastmc.dev",
+    "sources": "https://github.com/RemainingToast/BaritonePlus"
   },
   "license": "CC0-1.0",
-  "icon": "assets/altoclef/icon.png",
+  "icon": "assets/baritone-plus/baritone-plus.png",
   "environment": "*",
   "entrypoints": {
     "main": [
@@ -22,6 +22,6 @@
     ]
   },
   "mixins": [
-    "altoclef.mixins.json"
+    "baritone-plus.mixins.json"
   ]
 }
