+++ usage.md
@@ -1,6 +1,6 @@
 # Usage Guide
 
-Alto Clef has a variety of commands, settings and modes. This will give users an overview on how to use the bot.
+Bartione+ has a variety of commands, settings and modes. This will give users an overview on how to use the bot.
 
 Keep in mind this project is still in rapid development. A lot of features are placeholders and a work in progress.
